+++ src/Service/DataService.php
@@ -37,7 +37,7 @@ class DataService
     private $queueService;
 
     /**
-     * @var DataDownloaderService
+     * @var DataDownloader
      */
     private $dataDownloaderService;
 
@@ -59,7 +59,7 @@ class DataService
     public function __construct(
         QueueService $queueService,
         SolrService $solrService,
-        DataDownloaderService $downloaderService,
+        DataDownloader $downloaderService,
         array $indexableContentTypes,
         bool $retainDownloadedFiles,
         LoggerInterface $logger